requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
Pillow==10.2.0
googletrans==3.1.0a0
tqdm==4.66.2
colorama==0.4.6 